task: General refactor & cleanup
