+++ platformio.ini
@@ -15,3 +15,4 @@ build_flags =
   -D PIN_SCL_DISP=7 # GPIO07
   -D SERIAL_BAUD=115200
   '-D __FW_VERSION__="0.1.5"'
+  '-D __BUILD_DATE__=20240510'
